+++ index.js
@@ -1,5 +1,6 @@
 var pumpify = require('pumpify')
 var through = require('through2')
+var formatData = require('format-data')
 var events = require('events')
 var util = require('util')
 var uuid = require('cuid')
@@ -47,6 +48,21 @@ Dat.prototype.createImportStream = function (opts) {
 
   return pumpify(parseInputStream(opts), transform, writeStream)
 }
+
+
+Dat.prototype.createExportStream = function (opts) {
+  var parseOutput = through.obj(function (data, enc, next) {
+    debug('exporting through data', data)
+    if (data.content === 'row') {
+      var row = data.value
+      row.key = data.key
+      return next(null, row)
+    }
+  })
+
+  return pumpify(this.db.createReadStream(opts), parseOutput, formatData(opts.format))
+}
+
 
 Dat.prototype.createFileWriteStream = function (key, opts) {
   if (!opts.dataset) throw new Error('Error: opts.dataset required.')
